requirements.txt
colorlog
Flask
PyYAML
requests
